refactor(cockpit, web): extract useIsCoarsePointer + fix clamp docstring
Two follow-ups from review of #1236.

1. Move `useIsCoarsePointer` out of `CockpitView.tsx` into its own
   `hooks/useIsCoarsePointer.ts`. The hook duplicates the `(pointer:
   coarse)` matchMedia tracking already done by `useMobileKeyboard`'s
   `isMobile` field; extracting it makes the shared bit reusable so a
   future PR can rebase `useMobileKeyboard` onto it without dragging
   the keyboard listeners into every coarse / fine consumer. No
   behaviour change.

2. Fix the `isQueuedPromptLong` docstring in `queuedPromptsLayout.ts`:
   the implementation triggers on `lineCount >= 3` (i.e. 3+ lines / 2+
   newlines), but the comment claimed "3+ newlines" which is one
   off. The PR body and test descriptions already had the right
   threshold; this only updates the helper's own doc.
